validation fix

+++ sorting_validation.c
@@ -6,7 +6,7 @@
  * @array: The array to be printed
  * @size: Number of elements in @array
  */
-void sorting_validation(int *array, size_t size)
+void sorting_validation(int *array, int size)
 {
 
         if (!array || size < 2)
